fix lazy import for OpenAI module
